# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.ErrorResponse:
    description: Стандартный ответ с ошибкой
    properties:
      error:
        example: Authorization header required
        type: string
    type: object
...
  handlers.MaterialsResponse:
    description: Ответ с результатами поиска материалов
    properties:
      hasMore:
        example: true
        type: boolean
      limit:
        example: 20
        type: integer
      materials:
        items:
          $ref: '#/definitions/models.CatalogMaterial'
        type: array
      page:
        example: 1
        type: integer
      total:
        example: 150
        type: integer
    type: object
  handlers.ProfileResponse:
    description: Ответ с данными профиля пользователя
    properties:
      avatarUrl:
        example: https://example.com/avatars/123.jpg
        type: string
      createdAt:
        example: "2023-01-15T10:30:00Z"
        type: string
      email:
        example: user@example.com
        type: string
      fullName:
        example: Иван Иванов
        type: string
      id:
        example: 123
        type: integer
      isVerified:
        example: true
        type: boolean
      role:
        example: teacher
        type: string
      specializations:
        example:
        - math
        - physics
        items:
          type: string
        type: array
      updatedAt:
        example: "2023-01-15T10:30:00Z"
        type: string
    type: object
...
  handlers.SubjectsResponse:
    description: Ответ со списком предметов
    properties:
      subjects:
        items:
          $ref: '#/definitions/models.Subject'
        type: array
    type: object
...
  handlers.TeachersResponse:
    description: Ответ с результатами поиска преподавателей
    properties:
      teachers:
        items:
          $ref: '#/definitions/models.Teacher'
        type: array
    type: object
...
  handlers.UpdateProfileRequest:
    description: Запрос на обновление профиля
    properties:
      fullName:
        example: Иван Иванов
        type: string
      specializations:
        example:
        - math
        - physics
        items:
          type: string
        type: array
    type: object
  handlers.UpdateProfileResponse:
    description: Ответ на обновление профиля
    properties:
      data:
        $ref: '#/definitions/handlers.UpdateProfileRequest'
      message:
        example: Profile updated successfully
        type: string
      userID:
        example: 123
        type: integer
    type: object
  handlers.UploadAvatarResponse:
    description: Ответ на загрузку аватара
    properties:
      avatarUrl:
        example: https://example.com/avatars/123.jpg
        type: string
      message:
        example: Avatar uploaded successfully
        type: string
    type: object
...
  models.Author:
    description: Автор материала
    properties:
      id:
        example: 1
        type: integer
      name:
        example: Иван Иванов
        type: string
    type: object
...
  models.CatalogMaterial:
    description: Материал в каталоге
    properties:
      author:
        $ref: '#/definitions/models.Author'
      duration:
        example: 120
        type: integer
      id:
        example: 1
        type: integer
      level:
        example: beginner
        type: string
      rating:
        example: 4.8
        type: number
      studentsCount:
        example: 150
        type: integer
      subject:
        example: math
        type: string
      thumbnailUrl:
        example: https://example.com/thumbnail.jpg
        type: string
      title:
        example: Основы алгебры
        type: string
    type: object
...
  models.Subject:
    description: Учебный предмет
    properties:
      icon:
        example: "\U0001F4D0"
        type: string
      id:
        example: math
        type: string
      name:
        example: Математика
        type: string
    type: object
  models.Teacher:
    description: Преподаватель в каталоге
    properties:
      avatarUrl:
        example: https://example.com/avatar.jpg
        type: string
      id:
        example: 1
        type: integer
      materialsCount:
        example: 25
        type: integer
      name:
        example: Мария Петрова
        type: string
      rating:
        example: 4.9
        type: number
      specializations:
        example:
        - math
        - physics
        items:
          type: string
        type: array
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API для образовательной платформы Пайдея
  title: Paydeya Education Platform API
  version: "1.0"
paths:
...
  /catalog/materials:
    get:
      consumes:
      - application/json
      description: Возвращает материалы с фильтрацией и пагинацией
      parameters:
      - description: Поисковый запрос
        in: query
        name: search
        type: string
      - description: Фильтр по предмету
        in: query
        name: subject
        type: string
      - description: Фильтр по уровню сложности
        in: query
        name: level
        type: string
      - default: 1
        description: Номер страницы
        in: query
        name: page
        type: integer
      - default: 20
        description: Количество материалов на странице
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список материалов
          schema:
            $ref: '#/definitions/handlers.MaterialsResponse'
        "400":
          description: Неверные параметры запроса
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Поиск материалов в каталоге
      tags:
      - catalog
  /catalog/subjects:
    get:
      consumes:
      - application/json
      description: Возвращает все доступные учебные предметы
      produces:
      - application/json
      responses:
        "200":
          description: Список предметов
          schema:
            $ref: '#/definitions/handlers.SubjectsResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Получить список предметов
      tags:
      - catalog
  /catalog/teachers:
    get:
      consumes:
      - application/json
      description: Возвращает преподавателей с фильтрацией
      parameters:
      - description: Поисковый запрос
        in: query
        name: search
        type: string
      - description: Фильтр по предмету
        in: query
        name: subject
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список преподавателей
          schema:
            $ref: '#/definitions/handlers.TeachersResponse'
        "400":
          description: Неверные параметры запроса
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Поиск преподавателей
      tags:
      - catalog
...
  /profile:
    get:
      consumes:
      - application/json
      description: Возвращает данные профиля текущего пользователя
      produces:
      - application/json
      responses:
        "200":
          description: Данные профиля
          schema:
            $ref: '#/definitions/handlers.ProfileResponse'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Получить профиль пользователя
      tags:
      - profile
    put:
      consumes:
      - application/json
      description: Обновляет данные профиля пользователя
      parameters:
      - description: Данные для обновления
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Профиль обновлен
          schema:
            $ref: '#/definitions/handlers.UpdateProfileResponse'
        "400":
          description: Неверные данные
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Обновить профиль
      tags:
      - profile
  /profile/avatar:
    post:
      consumes:
      - multipart/form-data
      description: Загружает аватар пользователя
      parameters:
      - description: Файл аватара (макс. 5MB)
        in: formData
        name: avatar
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Аватар загружен
          schema:
            $ref: '#/definitions/handlers.UploadAvatarResponse'
        "400":
          description: Неверный файл
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Загрузить аватар
      tags:
      - profile
...
securityDefinitions:
  ApiKeyAuth:
    description: 'Введите: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...